requirements.txt
boto3==1.18.28
botocore==1.21.28
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
jmespath==0.10.0
mysql-connector==2.2.9
psycopg2==2.7.7
python-dateutil==2.8.2
PyYAML==5.4.1
pyyml==0.0.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
